+++ project/src/components/logo/logo.tsx
@@ -3,7 +3,7 @@ import {Link} from 'react-router-dom';
 function Logo(): JSX.Element {
   return (
     <div className="logo">
-      <Link to='/' className="logo__link">
+      <Link to="/" className="logo__link">
         <span className="logo__letter logo__letter--1">W</span>
         <span className="logo__letter logo__letter--2">T</span>
         <span className="logo__letter logo__letter--3">W</span>
