+++ project/src/components/pages/main-screen/main-screen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Film, Films } from '../../../mocks/films';
+import { Film, Films } from '../../../types/types';
 import FilmsList from './films-list/films-list';
 import FilmPromo from './film-promo/film-promo';
 import PageFooter from '../../page-footer/page-footer';
