+++ project/src/components/pages/movie-screen/film-control/film-control.tsx
@@ -1,5 +1,5 @@
 import { Link, useNavigate } from 'react-router-dom';
-import { Film } from '../../../../mocks/films';
+import { Film } from '../../../../types/types';
 import { AppRoute } from '../../../../const';
 
 type FilmControlProps = {
