+++ project/src/mocks/films.ts
@@ -1,24 +1,4 @@
-type Film = {
-  name: string,
-  posterImage: string,
-  previewImage: string,
-  backgroundImage: string,
-  backgroundColor: string,
-  description: string,
-  rating: number,
-  scoresCount: number,
-  director: string,
-  starring: string[],
-  runTime: number,
-  genre: string,
-  released: number,
-  id: number,
-  isFavorite: boolean,
-  videoLink: string,
-  previewVideoLink: string,
-}
-
-type Films = Film[];
+import { Film, Films } from '../types/types';
 
 const PROMO_FILM: Film =
 {
@@ -510,4 +490,4 @@ const FILMS: Films =
 ];
 
 export { PROMO_FILM, FILMS };
-export type { Film, Films };
+
