+++ project/src/types/types.ts
@@ -1,0 +1,23 @@
+type Film = {
+  name: string,
+  posterImage: string,
+  previewImage: string,
+  backgroundImage: string,
+  backgroundColor: string,
+  description: string,
+  rating: number,
+  scoresCount: number,
+  director: string,
+  starring: string[],
+  runTime: number,
+  genre: string,
+  released: number,
+  id: number,
+  isFavorite: boolean,
+  videoLink: string,
+  previewVideoLink: string,
+}
+
+type Films = Film[];
+
+export type { Film, Films };
